separate duties
